+++ pkg/channel/subprocess/procattr_other.go
@@ -1,0 +1,7 @@
+//go:build !unix
+
+package subprocess
+
+import "syscall"
+
+func procAttr() *syscall.SysProcAttr { return nil }